QL_Phong+Momo+QuenMK+DoiMK

+++ WebApplication1/Models/DatPhongInfo.cs
@@ -14,8 +14,8 @@ namespace WebApplication1.Models
         public DateTime NgayNhan { get; set; }
         public DateTime NgayTra { get; set; }
         public string TinhTrang { get; set; }
-        public float DonGia { get; set; }
-        public float DatCoc { get; set; }
+        public decimal DonGia { get; set; }
+        public decimal DatCoc { get; set; }
         public List<string> SelectedServices { get; set; }
  
     }
